chore: clean up swappercli some
