+++ packages/swapper/sample.env
@@ -1,5 +1,12 @@
-UNCHAINED_HTTPS_API=https://api.ethereum.shapeshift.com
-UNCHAINED_WS_API=wss://api.ethereum.shapeshift.com
 ETH_NODE_URL=eth node url goes here
 DEVICE_ID=device123
-MNEMONIC=salon adapt foil saddle orient make page zero cheese marble test catalog
+MNEMONIC=all all all all all all all all all all all all
+
+UNCHAINED_ETH_HTTPS_API=https://dev-api.ethereum.shapeshift.com
+UNCHAINED_ETH_WS_API=wss://dev-api.ethereum.shapeshift.com
+
+UNCHAINED_BTC_HTTPS_API=https://dev-api.bitcoin.shapeshift.com
+UNCHAINED_BTC_WS_API=wss://dev-api.bitcoin.shapeshift.com
+
+UNCHAINED_RUNE_HTTPS_API=https://dev-api.thorchain.shapeshift.com
+UNCHAINED_RUNE_WS_API=wss://dev-api.thorchain.shapeshift.com
